// migrations/snapshots/2faf416d736b01ccd8f82d6609a5bdbc0819fcf9239fbdc9c175585fa49ef11e/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly BoatPlacement: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly sessionId: CodecTypes['pg/text@1']['output'];
      readonly teamId: CodecTypes['pg/text@1']['output'];
      readonly exerciseId: CodecTypes['pg/text@1']['output'];
      readonly ownerId: CodecTypes['pg/text@1']['output'];
      readonly shipId: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Evaluation: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly sessionId: CodecTypes['pg/text@1']['output'];
      readonly teamId: CodecTypes['pg/text@1']['output'];
      readonly evaluatorId: CodecTypes['pg/text@1']['output'];
      readonly exerciseId: CodecTypes['pg/text@1']['output'];
      readonly repsObserved: CodecTypes['pg/int4@1']['output'];
      readonly note: CodecTypes['pg/int4@1']['output'];
      readonly isValidated: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly RefereeFleet: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly sessionId: CodecTypes['pg/text@1']['output'];
      readonly refereeId: CodecTypes['pg/text@1']['output'];
      readonly slot: CodecTypes['pg/int4@1']['output'];
      readonly status: 'PLACING' | 'LOCKED';
      readonly lockedAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly RefereeShip: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly fleetId: CodecTypes['pg/text@1']['output'];
      readonly size: CodecTypes['pg/int4@1']['output'];
      readonly orientation: CodecTypes['pg/text@1']['output'];
      readonly startTeamId: CodecTypes['pg/text@1']['output'];
      readonly startExerciseId: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Score: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly teamId: CodecTypes['pg/text@1']['output'];
      readonly points: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Session: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly wodType: 'PYRAMIDE_CLASSIQUE' | 'RELAIS_SPRINT' | 'TOUCHE_COULE_HYROX';
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly settings: CodecTypes['pg/json@1']['output'] | null;
      readonly raceEndedAt: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Shot: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly sessionId: CodecTypes['pg/text@1']['output'];
      readonly refereeId: CodecTypes['pg/text@1']['output'];
      readonly targetTeamId: CodecTypes['pg/text@1']['output'];
      readonly targetExerciseId: CodecTypes['pg/text@1']['output'];
      readonly evaluationId: CodecTypes['pg/text@1']['output'];
      readonly phase: 'DURING_WOD' | 'POST_WOD';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Team: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly sessionId: CodecTypes['pg/text@1']['output'];
    };
    readonly TeamMember: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly teamId: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/text@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly role: 'MASTER_ADMIN' | 'ADMIN' | 'GREFFIER' | 'STUDENT';
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly firstName: CodecTypes['pg/text@1']['output'] | null;
      readonly lastName: CodecTypes['pg/text@1']['output'] | null;
      readonly className: CodecTypes['pg/text@1']['output'] | null;
      readonly sex: 'M' | 'F' | null;
      readonly dateOfBirth: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly password: CodecTypes['pg/text@1']['output'] | null;
      readonly pinCode: CodecTypes['pg/text@1']['output'] | null;
      readonly profileVisibility: 'PUBLIC' | 'PRIVATE' | 'SECRET';
      readonly reliability: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly BoatPlacement: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly sessionId: CodecTypes['pg/text@1']['input'];
      readonly teamId: CodecTypes['pg/text@1']['input'];
      readonly exerciseId: CodecTypes['pg/text@1']['input'];
      readonly ownerId: CodecTypes['pg/text@1']['input'];
      readonly shipId: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Evaluation: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly sessionId: CodecTypes['pg/text@1']['input'];
      readonly teamId: CodecTypes['pg/text@1']['input'];
      readonly evaluatorId: CodecTypes['pg/text@1']['input'];
      readonly exerciseId: CodecTypes['pg/text@1']['input'];
      readonly repsObserved: CodecTypes['pg/int4@1']['input'];
      readonly note: CodecTypes['pg/int4@1']['input'];
      readonly isValidated: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly RefereeFleet: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly sessionId: CodecTypes['pg/text@1']['input'];
      readonly refereeId: CodecTypes['pg/text@1']['input'];
      readonly slot: CodecTypes['pg/int4@1']['input'];
      readonly status: 'PLACING' | 'LOCKED';
      readonly lockedAt: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly RefereeShip: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly fleetId: CodecTypes['pg/text@1']['input'];
      readonly size: CodecTypes['pg/int4@1']['input'];
      readonly orientation: CodecTypes['pg/text@1']['input'];
      readonly startTeamId: CodecTypes['pg/text@1']['input'];
      readonly startExerciseId: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Score: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly teamId: CodecTypes['pg/text@1']['input'];
      readonly points: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Session: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly wodType: 'PYRAMIDE_CLASSIQUE' | 'RELAIS_SPRINT' | 'TOUCHE_COULE_HYROX';
      readonly isActive: CodecTypes['pg/bool@1']['input'];
      readonly settings: CodecTypes['pg/json@1']['input'] | null;
      readonly raceEndedAt: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Shot: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly sessionId: CodecTypes['pg/text@1']['input'];
      readonly refereeId: CodecTypes['pg/text@1']['input'];
      readonly targetTeamId: CodecTypes['pg/text@1']['input'];
      readonly targetExerciseId: CodecTypes['pg/text@1']['input'];
      readonly evaluationId: CodecTypes['pg/text@1']['input'];
      readonly phase: 'DURING_WOD' | 'POST_WOD';
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Team: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly sessionId: CodecTypes['pg/text@1']['input'];
    };
    readonly TeamMember: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly teamId: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/text@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly role: 'MASTER_ADMIN' | 'ADMIN' | 'GREFFIER' | 'STUDENT';
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly firstName: CodecTypes['pg/text@1']['input'] | null;
      readonly lastName: CodecTypes['pg/text@1']['input'] | null;
      readonly className: CodecTypes['pg/text@1']['input'] | null;
      readonly sex: 'M' | 'F' | null;
      readonly dateOfBirth: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly password: CodecTypes['pg/text@1']['input'] | null;
      readonly pinCode: CodecTypes['pg/text@1']['input'] | null;
      readonly profileVisibility: 'PUBLIC' | 'PRIVATE' | 'SECRET';
      readonly reliability: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;